modularizes server and express
